# [ML] Functional tests - reduce setup module test execution time