## Summary

This PR stabilizes the setup module tests in cloud by reducing the execution time.

### Details

- When running the setup module tests on cloud, we sometimes hit the 4 minutes timeout for a job to finish, particularly for the `metrics_ui_hosts` and `metrics_ui_k8s` modules.
- An investigation showed that datafeeds with date histogram aggs don't stop on the last document but continue to run to the end date specified in the datafeed start request. With the small chunking config in some of these jobs and the test starting the datafeed with end date `now`, the jobs had to go through lots of empty time buckets, which cause a long job run time.
- While there are ideas how to optimize this kind of scenario on the backend side, we want to adjust the tests, so they're  stable on cloud.
- With this PR, the setup module tests don't start the datafeed until `now` anymore, but provide an end date close to the last record timestamp (rounded up the next 10 seconds for simpler numbers). This brings it closer to the way how module jobs are started from the UI, where we use the full source data time range by default.

### Test execution time changes

In the following list we can see how the test execution time changed on my local machine for each of the modules. It's overall mostly the same within 1 or 2 seconds, which I consider usual timing changes between test runs. But the `metrics_ui_hosts` and `metrics_ui_k8s` test execution time went down from > 1.5 minutes (which easily took more than 5 minutes on cloud) to less than 10 seconds.

module id | old time in ms | new time in ms
---|---|---
sample_data_weblogs | 5953 | 5615
apache_ecs | 22931 | 22390
apm_transaction | 4565 | 4317
logs_ui_analysis | 10642 | 9745
logs_ui_categories | 16731 | 15729
nginx_ecs | 8221 | 8215
sample_data_ecommerce | 9031 | 9183
siem_auditbeat_auth | 3999 | 4034
siem_packetbeat | 8365 | 8503
uptime_heartbeat | 4098 | 4015
auditbeat_process_hosts_ecs | 8276 | 7694
security_linux | 9965 | 10001
security_windows | 12398 | 12792
metricbeat_system_ecs | 5923 | 6896
metrics_ui_hosts | 90695 | 7627
metrics_ui_k8s | 101119 | 8163
siem_cloudtrail | 8659 | 8697
siem_winlogbeat | 19233 | 17461
siem_winlogbeat_auth | 4067 | 3679
apache_data_stream | 23737 | 22542
nginx_data_stream | 8666 | 8736